Cleaned up more, prepared for the thing
The thing is, by default the trend line should at first only show the
histogram (number of docs per month). When a month is chosen, the first
keyword should be selected. Whenever a keyword is selected, its
frequency is displayed as a line graph over the histogram. Consider
adding multiple selection of keywords after that is in place.